Extract `StoreSpy` into its own file.
